backend_fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
# Pin pydantic to a version compatible with Python 3.9.0 to avoid OpenAPI schema errors
pydantic==2.9.2
pydantic-core==2.23.4
# Psycopg 3 Postgres driver + connection pool (pip package is "psycopg-pool")
psycopg==3.2.3
psycopg-pool==3.2.2

# Auth & security
python-jose==3.3.0
passlib[bcrypt]==1.7.4

# Uploads and testing
python-multipart==0.0.9
httpx==0.27.2

# Data processing for calificaciones upload/export
pandas==2.2.2
openpyxl==3.1.5
